Indexing tests (#1979)
* fix missing insertion into lancedb index

* successfully running codebaseindex test

* await

* fix typo

* clean up indexing tests

* expectPlan

+++ core/test/testUtils/testDir.ts
@@ -11,17 +11,22 @@ export function setUpTestDir() {
 }
 
 export function tearDownTestDir() {
-  fs.rmSync(TEST_DIR, { recursive: true });
+  if (fs.existsSync(TEST_DIR)) {
+    fs.rmSync(TEST_DIR, { recursive: true });
+  }
 }
 
 export function addToTestDir(paths: (string | string[])[]) {
   for (const p of paths) {
+    const filepath = path.join(TEST_DIR, Array.isArray(p) ? p[0] : p);
+    fs.mkdirSync(path.dirname(filepath), { recursive: true });
+
     if (Array.isArray(p)) {
-      fs.writeFileSync(path.join(TEST_DIR, p[0]), p[1]);
+      fs.writeFileSync(filepath, p[1]);
     } else if (p.endsWith("/")) {
-      fs.mkdirSync(path.join(TEST_DIR, p), { recursive: true });
+      fs.mkdirSync(filepath, { recursive: true });
     } else {
-      fs.writeFileSync(path.join(TEST_DIR, p), "");
+      fs.writeFileSync(filepath, "");
     }
   }
 }
